Rename from create text to text creation page
